api/requirements.txt
flask==2.0.1
werkzeug<2.1
redis==4.0.2
mysql-connector-python==8.0.27
python-dotenv==0.19.1
pybase62